moves ci logic to matrix
