+++ rule_option.go
@@ -1,0 +1,8 @@
+// Copyright The OWASP Coraza contributors
+// SPDX-License-Identifier: Apache-2.0
+
+//go:build !coraza.rule.multiphase_evaluation
+
+package main
+
+const multiphaseEvaluation = false
